+++ Outerrain/Include/vec.h
@@ -13,13 +13,13 @@ struct Vector4;
 struct Vector2
 {
 public:
-	double x, y;
+	float x, y;
 
 	Vector2() : x(0.0), y(0.0) {}
-	Vector2(double n) : x(n), y(n) {}
-	Vector2(double x, double y) : x(x), y(y) {}
+	Vector2(float n) : x(n), y(n) {}
+	Vector2(float x, float y) : x(x), y(y) {}
 
-	Vector2 operator*(const double& k) const
+	Vector2 operator*(const float& k) const
 	{
 		return Vector2(x * k, y * k);
 	}
@@ -35,7 +35,7 @@ public:
 	{
 		return Vector2(x + u.x, y + u.y);
 	}
-	double operator[](int i)
+	float operator[](int i)
 	{
 		if (i == 0)
 			return x;
@@ -48,17 +48,17 @@ inline std::ostream& operator<<(std::ostream& stream, const Vector2& u)
 	stream << "(" << u.x << ", " << u.y << ");";
 	return stream;
 }
-inline double Dot(const Vector2& u, const Vector2& v)
+inline float Dot(const Vector2& u, const Vector2& v)
 {
 	return u.x * v.x + u.y * v.y;
 }
-inline double Magnitude(const Vector2& u)
+inline float Magnitude(const Vector2& u)
 {
 	return sqrt(u.x * u.x + u.y * u.y);
 }
 inline Vector2 Normalize(const Vector2& v)
 {
-	double kk = 1 / Magnitude(v);
+	float kk = 1 / Magnitude(v);
 	return v * kk;
 }
 inline Vector2 Center(const Vector2& a, const Vector2& b)
@@ -75,13 +75,13 @@ inline Vector2 operator-(const Vector2& v)
 struct Vector3
 {
 public:
-	double x, y, z;
+	float x, y, z;
 
 	explicit Vector3() : x(0.0), y(0.0), z(0.0) { }
-	explicit Vector3(double n) : x(n), y(n), z(n) {}
-	explicit Vector3(double x, double y, double z) : x(x), y(y), z(z) {}
+	explicit Vector3(float n) : x(n), y(n), z(n) {}
+	explicit Vector3(float x, float y, float z) : x(x), y(y), z(z) {}
 
-	Vector3 operator*(const double& k) const
+	Vector3 operator*(const float& k) const
 	{
 		return Vector3(x * k, y * k, z * k);
 	}
@@ -97,7 +97,7 @@ public:
 	{
 		return Vector3(x + u.x, y + u.y, z + u.z);
 	}
-	double operator[](int i)
+	float operator[](int i)
 	{
 		if (i == 0)
 			return x;
@@ -119,17 +119,17 @@ inline Vector3 Cross(const Vector3& u, const Vector3& v)
 		(u.z * v.x) - (u.x * v.z),
 		(u.x * v.y) - (u.y * v.x));
 }
-inline double Dot(const Vector3& u, const Vector3& v)
+inline float Dot(const Vector3& u, const Vector3& v)
 {
 	return u.x * v.x + u.y * v.y + u.z * v.z;
 }
-inline double Magnitude(const Vector3& u)
+inline float Magnitude(const Vector3& u)
 {
 	return sqrt(u.x * u.x + u.y * u.y + u.z * u.z);
 }
 inline Vector3 Normalize(const Vector3& v)
 {
-	double kk = 1 / Magnitude(v);
+	float kk = 1 / Magnitude(v);
 	return v * kk;
 }
 inline Vector3 Center(const Vector3& a, const Vector3& b)
@@ -146,13 +146,13 @@ inline Vector3 operator-(const Vector3& v)
 struct Vector4
 {
 public:
-	double x, y, z, w;
+	float x, y, z, w;
 
 	Vector4() : x(0.0), y(0.0), z(0.0), w(0.0) {}
-	Vector4(double n) : x(n), y(n), z(n), w(n) {}
-	Vector4(double x, double y, double z, double w) : x(x), y(y), z(z), w(w) {}
+	Vector4(float n) : x(n), y(n), z(n), w(n) {}
+	Vector4(float x, float y, float z, float w) : x(x), y(y), z(z), w(w) {}
 
-	Vector4 operator*(const double& k) const
+	Vector4 operator*(const float& k) const
 	{
 		return Vector4(x * k, y * k, z * k, w * k);
 	}
@@ -168,7 +168,7 @@ public:
 	{
 		return Vector4(x + u.x, y + u.y, z + u.z, w + u.w);
 	}
-	double operator[](int i)
+	float operator[](int i)
 	{
 		if (i == 0)
 			return x;
@@ -180,13 +180,13 @@ public:
 	}
 	friend std::ostream& operator<<(std::ostream& stream, const Vector4& u);
 };
-inline double Magnitude(const Vector4& u)
+inline float Magnitude(const Vector4& u)
 {
 	return sqrt(u.x * u.x + u.y * u.y + u.z * u.z + u.w * u.w);
 }
 inline Vector4 Normalize(const Vector4& v)
 {
-	double kk = 1 / Magnitude(v);
+	float kk = 1 / Magnitude(v);
 	return v * kk;
 }
 inline std::ostream& operator<<(std::ostream& stream, const Vector4& u)
